+++ bin/main.rb
@@ -1,11 +1,13 @@
-require_relative '../lib/js-linter.rb'
+# frozen_string_literal: true
 
+require_relative '../lib/js_linter'
+# top-level class documentation
 class JSLinter
   def initialize(file)
     puts 'Start JS Linter initializing... '
-    js_linter_controller = JSLinterControl.new(file)
+    JSLinterControl.new(file)
   end
 end
 
 file = ARGV[0]
-js_linter = JSLinter.new(file)
+JSLinter.new(file)
