+++ lib/checker.rb
@@ -1,3 +1,9 @@
+# frozen_string_literal: true
+
+require_relative 'file_reader'
+# top-level class documentation
 class Checker
-  
+  def initialize
+    puts arr_file
+  end
 end
